@@ -9,7 +9,7 @@ import org.schabi.newpipe.R
 import org.schabi.newpipe.extractor.channel.ChannelInfoItem
 import org.schabi.newpipe.util.Localization
 import org.schabi.newpipe.util.OnClickGesture
-import org.schabi.newpipe.util.image.PicassoHelper
+import org.schabi.newpipe.util.image.CoilHelper
 
 class ChannelItem(
     private val infoItem: ChannelInfoItem,
@@ -39,7 +39,7 @@ class ChannelItem(
             itemChannelDescriptionView.text = infoItem.description
         }
 
-        PicassoHelper.loadAvatar(infoItem.thumbnails).into(itemThumbnailView)
+        CoilHelper.loadAvatar(itemThumbnailView, infoItem.thumbnails)
 
         gesturesListener?.run {
             viewHolder.root.setOnClickListener { selected(infoItem) }
