@@ -6,8 +6,8 @@ import android.view.MotionEvent;
 import android.view.View;
 
 import org.schabi.newpipe.util.Localization;
-import org.schabi.newpipe.util.image.PicassoHelper;
 import org.schabi.newpipe.util.ServiceHelper;
+import org.schabi.newpipe.util.image.CoilHelper;
 
 public class PlayQueueItemBuilder {
     private static final String TAG = PlayQueueItemBuilder.class.toString();
@@ -33,7 +33,7 @@ public class PlayQueueItemBuilder {
             holder.itemDurationView.setVisibility(View.GONE);
         }
 
-        PicassoHelper.loadThumbnail(item.getThumbnails()).into(holder.itemThumbnailView);
+        CoilHelper.INSTANCE.loadThumbnail(holder.itemThumbnailView, item.getThumbnails());
 
         holder.itemRoot.setOnClickListener(view -> {
             if (onItemClickListener != null) {
