@@ -13,10 +13,9 @@ import org.schabi.newpipe.extractor.NewPipe;
 import org.schabi.newpipe.extractor.localization.ContentCountry;
 import org.schabi.newpipe.extractor.localization.Localization;
 import org.schabi.newpipe.util.image.ImageStrategy;
-import org.schabi.newpipe.util.image.PicassoHelper;
 import org.schabi.newpipe.util.image.PreferredImageQuality;
 
-import java.io.IOException;
+import coil.Coil;
 
 public class ContentSettingsFragment extends BasePreferenceFragment {
     private String youtubeRestrictedModeEnabledKey;
@@ -42,14 +41,13 @@ public class ContentSettingsFragment extends BasePreferenceFragment {
                 (preference, newValue) -> {
                     ImageStrategy.setPreferredImageQuality(PreferredImageQuality
                             .fromPreferenceKey(requireContext(), (String) newValue));
-                    try {
-                        PicassoHelper.clearCache(preference.getContext());
-                        Toast.makeText(preference.getContext(),
-                                R.string.thumbnail_cache_wipe_complete_notice, Toast.LENGTH_SHORT)
-                                .show();
-                    } catch (final IOException e) {
-                        Log.e(TAG, "Unable to clear Picasso cache", e);
-                    }
+                    final var loader = Coil.imageLoader(preference.getContext());
+                    loader.getMemoryCache().clear();
+                    loader.getDiskCache().clear();
+                    Toast.makeText(preference.getContext(),
+                            R.string.thumbnail_cache_wipe_complete_notice, Toast.LENGTH_SHORT)
+                            .show();
+
                     return true;
                 });
     }
